backend/pyproject.toml
[tool.poetry]
name = "sidequest-overlay-backend"
version = "1.0.0"
description = "FastAPI backend for SideQuest real-time GPS overlay with AI-powered narrative"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "*.py"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = ">=0.115.0"
uvicorn = {extras = ["standard"], version = ">=0.32.0"}
websockets = ">=14.0"
pydantic = ">=2.10.0"
openai = ">=1.54.0"
python-dotenv = ">=1.0.0"
requests = "^2.32.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn main:app --reload --port 8787"

